pyproject.toml
[tool.poetry]
name = "crawl_repos"
version = "0.1.0"
description = ""
authors = ["Takayuki Zukawa <zk.takayuki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
gql = {extras = ["all"], version = "^3.0.0"}
python-dotenv = "^0.19.2"
pandas = "^1.4.0"
requests = "^2.27.1"
requests-oauthlib = "^1.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
